move touch proxy to emGUIGlue file
